feat(bench_saturate): add CLI commands, main with system probe, and headless mode
Phase 4-5: Complete App 17 implementation.

- 8 CLI commands: /saturate start|stop|status|report|export|list|probe|help
  following bench_perf CommandRegistration<T> pattern
- System probe: CPU cores (P/E split on macOS), cache topology, total memory
- Headless mode: --headless <preset> [--format json|csv] [--output <path>]
  with timed wait loop and result output
- Rate cap (100M msg/s) prevents uint32_t overflow in ramp

26 unit tests pass (12 messages + 14 collector math). Headless mode
produces valid JSON/CSV with throughput, latency, and drop metrics.

Closes: #313

+++ apps/bench_saturate/actors/saturate_receiver_actor.hpp
@@ -153,7 +153,7 @@ class SaturateReceiverActor : public EventBasedActor {
 
     void send_drop_report() {
         DropReportPayload report;
-        report.receiver_id = static_cast<uint32_t>(id().value());
+        report.receiver_id = id().value();
         report.total_received = received_count_.load();
         report.total_dropped = dropped_count_.load();
         context()->send(collector_addr_, make_msg(DropReportTag, report.encode()));
